Split engine into modules; added menu, mouse
common.js is now separated into game.js, camera.js, input.js, loader.js,
menu.js and text.js. menu.js and text.js are empty for now, but will
ultimately play a similar role to camera.js, i.e. holding the logic for
a specific game view. Currently all rendering is done in game.js and I
believe it should stay that way for the forseeable future.

The most obvious front-facing change is the addition of a menu overlay
to the map mode. A quick todo list of the work that still needs to be
done:
* Buttons aren't hooked up to anything yet
* Relevant logic needs to moved into the menu.js module
* Camera doesn't recognize the beginning of the menu as the end of the map
for scrolling purposes...

input.js holds the old Keyboard class and also the new Mouse class,
which returns click coordinates to the Game class that can then be used
by any other game sub-module.